Kill Controller\OG class
