+++ src/Og.php
@@ -7,15 +7,148 @@
 
 namespace Drupal\og;
 
-use Drupal\Core\Config\Entity\ConfigEntityInterface;
-use Drupal\Core\Entity\ContentEntityInterface;
-use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Field\FieldConfigInterface;
+use Drupal\field\Entity\FieldConfig;
+use Drupal\field\Entity\FieldStorageConfig;
+use Drupal\Core\Entity\Entity\EntityFormDisplay;
 
 /**
  * A static helper class for OG.
  */
 class Og {
+
+  /**
+   * Static cache for groups per entity.
+   *
+   * @var array
+   */
+  protected static $entityGroupCache = [];
+
+  /**
+   * Create an organic groups field in a bundle.
+   *
+   * @param $field_name
+   *   The field name.
+   * @param $entity_type
+   *   The entity type.
+   * @param $bundle
+   *   The bundle name.
+   */
+  public static function createField($field_name, $entity_type, $bundle) {
+    $og_field = static::fieldInfo($field_name)
+      ->setEntityType($entity_type)
+      ->setBundle($bundle);
+
+    if (!FieldStorageConfig::loadByName($entity_type, $field_name)) {
+      $og_field->fieldDefinition()->save();
+    }
+
+    // Allow overriding the field name.
+    // todo: ask if we need this.
+//    $og_field['field']['field_name'] = $field_name;
+//    if (empty($field)) {
+//      $og_field['field']->save();
+//    }
+
+    if (!FieldConfig::loadByName($entity_type, $bundle, $field_name)) {
+      $og_field->instanceDefinition()->save();
+      // Clear the entity property info cache, as OG fields might add different
+      // entity property info.
+      static::invalidateCache();
+    }
+
+    // Add the field to the form display manager.
+    $displayForm = EntityFormDisplay::load($entity_type . '.' . $bundle . '.default');
+    if (!$displayForm->getComponent($field_name) && $widgetDefinition = $og_field->widgetDefinition()) {
+      $displayForm->setComponent($field_name, $widgetDefinition);
+      $displayForm->save();
+    }
+
+    // Define the view mode for the field.
+    if ($fieldViewModes = $og_field->viewModesDefinition()) {
+      $prefix = $entity_type . '.' . $bundle . '.';
+      $viewModes = entity_load_multiple('entity_view_display', array_keys($fieldViewModes));
+
+      foreach ($viewModes as $key => $viewMode) {
+        $viewMode->setComponent($field_name, $fieldViewModes[$prefix . $key])->save();
+      }
+    }
+  }
+
+  /**
+   * Gets the groups an entity is associated with.
+   *
+   * @param $entity_type
+   *   The entity type. Defaults to 'user'
+   * @param $entity_id
+   *   (optional) The entity ID.
+   * @param $states
+   *   (optional) Array with the state to return. Defaults to active.
+   * @param $field_name
+   *   (optional) The field name associated with the group.
+   *
+   * @return array
+   *  An array with the group's entity type as the key, and array - keyed by
+   *  the OG membership ID and the group ID as the value. If nothing found,
+   *  then an empty array.
+   */
+  public static function getEntityGroups($entity_type = 'user', $entity_id = NULL, $states = [OG_STATE_ACTIVE], $field_name = NULL) {
+    $cache = &drupal_static(__FUNCTION__, array());
+
+    if ($entity_type == 'user' && empty($entity_id)) {
+      $account = \Drupal::currentUser()->getAccount();
+      $entity_id = $account->id();
+    }
+
+    // Get a string identifier of the states, so we can retrieve it from cache.
+    if ($states) {
+      sort($states);
+      $state_identifier = implode(':', $states);
+    }
+    else {
+      $state_identifier = FALSE;
+    }
+
+    $identifier = [
+      $entity_type,
+      $entity_id,
+      $state_identifier,
+      $field_name,
+    ];
+
+    $identifier = implode(':', $identifier);
+    if (isset(static::$entityGroupCache[$identifier])) {
+      // Return cached values.
+      return static::$entityGroupCache[$identifier];
+    }
+
+    $cache[$identifier] = [];
+    $query = \Drupal::entityQuery('og_membership')
+      ->condition('entity_type', $entity_type)
+      ->condition('etid', $entity_id);
+
+    if ($states) {
+      $query->condition('state', $states, 'IN');
+    }
+
+    if ($field_name) {
+      $query->condition('field_name', $field_name);
+    }
+
+    $results = $query->execute();
+
+    /** @var OgMembership[] $memberships */
+    $memberships = \Drupal::entityManager()
+      ->getStorage('og_membership')
+      ->loadMultiple($results);
+
+    /** @var \Drupal\og\Entity\OgMembership $membership */
+    foreach ($memberships as $membership) {
+      static::$entityGroupCache[$identifier][$membership->getGroupType()][$membership->id()] = $membership->getGroup();
+    }
+
+    return static::$entityGroupCache[$identifier];
+  }
 
   /**
    * Check if the given entity is a group.
@@ -70,6 +203,81 @@ class Og {
   public static function groupManager() {
     // @todo store static reference for this?
     return \Drupal::service('og.group.manager');
+  }
+
+  /**
+   * Invalidate cache.
+   *
+   * @param $gids
+   *   Array with group IDs that their cache should be invalidated.
+   */
+  public static function invalidateCache($gids = array()) {
+    // @todo We should not be using drupal_static() review and remove.
+    // Reset static cache.
+    $caches = array(
+      'og_user_access',
+      'og_user_access_alter',
+      'og_role_permissions',
+      'og_get_user_roles',
+      'og_get_permissions',
+      'og_get_group_audience_fields',
+      'og_get_entity_groups',
+      'og_get_membership',
+      'og_get_field_og_membership_properties',
+      'og_get_user_roles',
+    );
+
+    foreach ($caches as $cache) {
+      drupal_static_reset($cache);
+    }
+
+    // @todo Consider using a reset() method.
+    static::$entityGroupCache = [];
+
+    // Invalidate the entity property cache.
+    \Drupal::entityManager()->clearCachedDefinitions();
+    \Drupal::entityManager()->clearCachedFieldDefinitions();
+
+    // Let other OG modules know we invalidate cache.
+    \Drupal::moduleHandler()->invokeAll('og_invalidate_cache', $gids);
+  }
+
+  /**
+   * Gets the storage manage for the OG membership entity.
+   *
+   * @return \Drupal\Core\Entity\EntityStorageInterface
+   */
+  public static function membershipStorage() {
+    return \Drupal::entityManager()->getStorage('og_membership');
+  }
+
+  /**
+   * Gets the default constructor parameters for OG membership.
+   */
+  public static function membershipDefault() {
+    return ['type' => 'og_membership_type_default'];
+  }
+
+  /**
+   * Get all the modules fields that can be assigned to fieldable entities.
+   *
+   * @param $field_name
+   *   The field name that was registered for the definition.
+   *
+   * @return OgFieldBase|bool
+   *   An array with the field and instance definitions, or FALSE if not.
+   *
+   * todo: pass the entity type and entity bundle to plugin definition.
+   */
+  protected static function fieldInfo($field_name = NULL) {
+    $config = \Drupal::service('plugin.manager.og.fields');
+    $fields_config = $config->getDefinitions();
+
+    if ($field_name) {
+      return isset($fields_config[$field_name]) ? $config->createInstance($field_name) : NULL;
+    }
+
+    return $fields_config;
   }
 
 }
